chore: fix reviewer comments
