+++ src/constant/index.ts
@@ -6,4 +6,4 @@ export * from "./node";
 export * from "./token";
 export * from "./walletProvider";
 export * from "./eip6963Provider";
-export * from "./chainConfig";
+export * from "./web3Config";
