+++ src/provider/metamask/MetaMask.ts
@@ -27,13 +27,6 @@ import { signTransactionWrapper } from "@carbon-sdk/util/provider";
 import { Eip6963Provider } from "../eip6963Provider";
 import { ARBITRUM_MAINNET, ARBITRUM_TESTNET, BSC_MAINNET, BSC_TESTNET, CARBON_EVM_DEVNET, CARBON_EVM_LOCALHOST, CARBON_EVM_MAINNET, CARBON_EVM_TESTNET, ETH_MAINNET, ETH_TESTNET, ChangeNetworkParam as MetaMaskChangeNetworkParam, OKC_MAINNET, OKC_TESTNET, POLYGON_MAINNET, POLYGON_TESTNET } from "../../constant";
 
-
-declare global{
-  interface WindowEventMap {
-    "eip6963:announceProvider": CustomEvent
-  }
-}
-
 export type EVMChain = EVMChainV2;
 
 type ChainContracts = {
