--- src/types.d.ts
@@ -1,34 +1,0 @@
-declare module "ledger-cosmos-js";
-interface EIP6963ProviderDetail {
-  info: EIP6963ProviderInfo;
-  provider: EIP1193Provider;
-}
-
-interface EIP6963ProviderInfo {
-  walletId: string;
-  uuid: string;
-  name: string;
-  icon: string;
-}
-
-interface EIP1193Provider {
-  isStatus?: boolean;
-  host?: string;
-  path?: string;
-  isMetaMask: boolean;
-  chainId: string;
-  sendAsync?: (request: { method: string, params?: Array<unknown> }, callback: (error: Error | null, response: unknown) => void) => void
-  send?: (request: { method: string, params?: Array<unknown> }, callback: (error: Error | null, response: unknown) => void) => void
-  request: (request: { method: string, params?: Array<unknown> }) => Promise<unknown>
-  on: (eventName: string, listener: (...args: unknown[]) => void) => any
-  _state?: {
-    isConnected?: boolean
-  }
-}
-
-type EIP6963AnnounceProviderEvent = {
-  detail:{
-    info: EIP6963ProviderInfo,
-    provider: EIP1193Provider
-  }
-}
