+++ src/types/index.d.ts
@@ -1,0 +1,6 @@
+
+declare module "ledger-cosmos-js";
+
+interface WindowEventMap {
+  "eip6963:announceProvider": CustomEvent
+}
